Separate dialogs in different files
