+++ src/renderer/src/pages/index/generation.dialog.tsx
@@ -1,0 +1,75 @@
+import { AnimeSong } from "@renderer/types/list.types";
+import { useState } from "react";
+import { Step, STEPS } from "./dialogs/dialog.constants";
+
+import ConfirmationDialog from "./dialogs/confirmation.dialog";
+import CopyrightProgressDialog from "./dialogs/copyright.dialog";
+import DownloadProgressDialog from "./dialogs/download.dialog";
+import RenderingHtmlDialog from "./dialogs/rendering.dialog";
+import CompletedDialog from "./dialogs/completed.dialog";
+
+type ProgressDialogProps = {
+  songs: AnimeSong[];
+  author: string;
+  outputDir: string;
+};
+
+export default function ProgressDialog({ songs, outputDir, author }: ProgressDialogProps) {
+  const [step, setStep] = useState<Step>(STEPS.CONFIRMATION);
+
+  const [unembeddableIds, setUnembeddableIds] = useState<Array<string>>([]);
+  const [failedIds, setFailedIds] = useState<Array<string>>([]);
+
+  const isDisabled = Boolean(!outputDir || songs.length === 0);
+
+  if (step === STEPS.CONFIRMATION) {
+    return <ConfirmationDialog isDisabled={isDisabled} setStep={setStep} />;
+  }
+
+  if (step === STEPS.COPYRIGHT) {
+    return (
+      <CopyrightProgressDialog
+        songs={songs}
+        setUnembeddableIds={setUnembeddableIds}
+        setStep={setStep}
+      />
+    );
+  }
+
+  if (step === STEPS.DOWNLOAD) {
+    return (
+      <DownloadProgressDialog
+        unembeddableIds={unembeddableIds}
+        setFailedIds={setFailedIds}
+        outputDir={outputDir}
+        setStep={setStep}
+      />
+    );
+  }
+
+  if (step === STEPS.RENDERING) {
+    return (
+      <RenderingHtmlDialog
+        author={author}
+        failedIds={failedIds}
+        outputDir={outputDir}
+        songs={songs}
+        unembeddableIds={unembeddableIds}
+        setStep={setStep}
+      />
+    );
+  }
+
+  if (step === STEPS.COMPLETED) {
+    return (
+      <CompletedDialog
+        outputDir={outputDir}
+        setFailedIds={setFailedIds}
+        setStep={setStep}
+        setUnembeddableIds={setUnembeddableIds}
+      />
+    );
+  }
+
+  return null;
+}
